chore: mild refactoring of files
